// frontend/src/typechain-types/MusicStore.ts
export declare namespace MusicStore {
  export type AlbumStruct = {
    index: PromiseOrValue<BigNumberish>;
    uid: PromiseOrValue<string>;
    title: PromiseOrValue<string>;
    price: PromiseOrValue<BigNumberish>;
    quantity: PromiseOrValue<BigNumberish>;
  };

  export type AlbumStructOutput = [
    BigNumber,
    string,
    string,
    BigNumber,
    BigNumber
  ] & {
    index: BigNumber;
    uid: string;
    title: string;
    price: BigNumber;
    quantity: BigNumber;
  };
}